test: separate evaluation api contract integration coverage
